Docker orchestration update
- API uses multi-layered Dockerfile
- All docker assets for api are in api/_docker
- Configured as API Platform is using cached repositories, however new caches need to be made base on this common build we will use across our web apps
- App now builds at entrypoint again - secrets should not be in builds and the nuxt build requires a cookie secret
- ./run bash file updates include memory limit imposed for build commands, stack and swarm command support

Although this update begins the optimisation process, it will not prevent the server from becoming overloaded. The primary culprit is the 'app' service. With these updates we move towards being able to run a single swarm instance which will respect the memory limits imposed in the docker-compose files. I think we can use swarms/stacks as described here: https://docs.docker.com/engine/reference/commandline/stack_deploy/ - note that some compose keys are not supported in docker stack commands (such as the build key). Not sure of the impact of this yet and possibilities to implement build servers to build each app before deploying. Possibly build and push each build to a build server and then deploy.

The 'admin' service has been removed from the docker compose files for the time being. Starting the admin is VERY slow and uses a huge amount of resources. Once a build server is implemented this should become easier.

These updates have only been implemented in Addiction PAT website so far and once optimisation is complete will need to be implemented in each web app currently running on the current production server.